dev_match/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
drf-spectacular==0.28.0
pytest==8.3.4
coverage==7.6.12